Pass unit/multi.tcl: Lua script timeouts (SCRIPT KILL / BUSY), thin CONFIG, and WATCH-on-expiry fix (#772)
* enable multi tests

* add missing files

* update

* add test

+++ src/main/java/com/github/fppt/jedismock/operations/strings/SetRange.java
@@ -30,7 +30,10 @@ public class SetRange extends AbstractRedisOperation {
                 .map(RMString::getStoredDataAsString)
                 .orElse("");
         String padding = "";
-        if (offset + value.length() > base().getProtoMaxBulkLen()) {
+        //(long) cast: offset is an int near Integer.MAX_VALUE, so a plain
+        //int+int could overflow to a negative, skip this guard, and then attempt
+        //a multi-GB allocation below.
+        if ((long) offset + value.length() > base().getProtoMaxBulkLen()) {
             return Response.error("ERR string exceeds maximum allowed size (proto-max-bulk-len)");
         }
         if (offset > oldValue.length()) {
